pyproject.toml
[tool.poetry]
name = "torchdaiv"
version = "1.1.0"
description = "PyTorch Wrapper for dAiv's lecture"
authors = ["dAiv <cnudaiv@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [
    {include = "torchdaiv", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.9"
torchvision = "^0.18.0"
tqdm = "^4.66.4"
rich = "^13.7.1"
scipy = "1.12.0"
gensim = "^4.3.2"
konlpy = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
